fix: Fix non-functional vim plugin
The vim plugin uses --expect and expects the key used to accept the
selection on the first line of output. Commit bcee1f4 "feat!: do not
check for expect before printing the argument of accept… (#625)" broke
that by not outputting anything if the selection was made using the
default "enter" key. We can workaround that by explicitly adding
"enter" to the --expect argument, so that it once again shows up in the
output.

Fixes skim-rs/skim.vim#25